requirements.txt
# ================================
# GAL-Dreamer 依赖列表
# ================================

# 核心框架
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0

# LLM providers
openai>=1.0.0
dashscope>=1.0.0  # 通义千问

# 配置管理
python-dotenv>=1.0.0
pyyaml>=6.0.0

# 数据处理
pydantic>=2.0.0

# HTTP请求
requests>=2.31.0

# 工具库
tqdm>=4.66.0

# 日志
loguru>=0.7.0

# 测试
pytest>=7.4.0

# 代码质量
black>=23.12.0
flake8>=7.0.0
